Changed double price to BigDecimal price

+++ backend/src/main/java/model/Orden.java
@@ -1,5 +1,6 @@
 package model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record Orden(String id, Pedido pedido, Plato plato, double price, OrdenEstado ordenEstado, LocalDate fecha, String Detalles){}
+public record Orden(String id, Pedido pedido, Plato plato, BigDecimal price, OrdenEstado ordenEstado, LocalDate fecha, String Detalles){}
